gmk: remove unneeded typecasts
These two vars used to be marked as bool by gm8x, but they should actually be u32. Now that's fixed we don't need to cast them to u32 anymore.

+++ src/gmk.rs
@@ -573,8 +573,8 @@ where
         result += writer.write_i32_le(tile.x)?;
         result += writer.write_i32_le(tile.y)?;
         result += writer.write_i32_le(tile.source_bg)?;
-        result += writer.write_u32_le(tile.tile_x as u32)?;
-        result += writer.write_u32_le(tile.tile_y as u32)?;
+        result += writer.write_u32_le(tile.tile_x)?;
+        result += writer.write_u32_le(tile.tile_y)?;
         result += writer.write_u32_le(tile.width)?;
         result += writer.write_u32_le(tile.height)?;
         result += writer.write_i32_le(tile.depth)?;
